# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Calendar Sharing App API
  description: API for a group-based calendar sharing application
  version: 1.0.0
  contact:
    name: Team

servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.calendar-share.app/api/v1
    description: Production server

paths:
  /auth/register:
    post:
      summary: User registration
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                name:
                  type: string
              required:
                - email
                - password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or email already exists
        '500':
          description: Server error

  /auth/login:
    post:
      summary: User login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT token
        '401':
          description: Invalid credentials
        '500':
          description: Server error

  /users/me:
    get:
      summary: Get current user profile
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    patch:
      summary: Update current user profile (name / email / password / memo)
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 50
                email:
                  type: string
                  format: email
                  description: メールアドレスを変更する場合のみ指定
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 100
                  description: パスワードを変更する場合のみ指定
                memo:
                  type: string
                  maxLength: 200
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '409':
          description: Email already in use by another account

    delete:
      summary: Delete current user account
      description: |
        ログイン中のユーザーのアカウントを削除する。
        グループオーナーの場合は削除不可（403）。
        削除後は全グループのメンバーシップも削除される。
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account deleted successfully
        '400':
          description: Cannot delete — user is a group owner
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /groups:
    get:
      summary: Get all groups the current user belongs to
      tags:
        - Groups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of groups with members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupWithMembers'
        '401':
          description: Unauthorized
        '500':
          description: Server error

    post:
      summary: Create a new group
      tags:
        - Groups
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupWithMembers'
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /groups/me:
    get:
      summary: Get user's group
      tags:
        - Groups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupWithMembers'
        '401':
          description: Unauthorized
        '404':
          description: Group not found
        '500':
          description: Server error

  /groups/join/{token}:
    get:
      summary: Get group info by invite token (preview before joining)
      tags:
        - Groups
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Group info preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPreview'
        '401':
          description: Unauthorized
        '404':
          description: Invalid invite token
        '500':
          description: Server error

    post:
      summary: Join a group via invite link
      tags:
        - Groups
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully joined group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
        '404':
          description: Invalid invite token
        '409':
          description: User already in this group
        '500':
          description: Server error

  /groups/{groupId}:
    delete:
      summary: Dissolve a group (owner only)
      description: |
        グループオーナーのみグループを解散できる。
        解散するとグループ内の全データ（メンバーシップ・カレンダー・イベント・シフト）が削除される。
        この操作は取り消せない。
      tags:
        - Groups
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Group dissolved successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not the group owner
        '404':
          description: Group not found
        '500':
          description: Server error

  /groups/{groupId}/members/{userId}:
    delete:
      summary: Remove a member from a group
      description: |
        グループオーナーは任意のメンバーを退会させることができる。
        メンバー本人も自分自身を退会させることができる（自己退会）。
        グループオーナー自身の退会は不可。
        退会後も招待リンクから再参加可能。
      tags:
        - Groups
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Member removed successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not group owner or not the member themselves / owner cannot remove themselves
        '404':
          description: Group or member not found
        '500':
          description: Server error

  /calendars:
    get:
      summary: Get calendars in user's groups (optionally filtered by groupId)
      tags:
        - Calendars
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter calendars by specific group (must be a group the user belongs to)
      responses:
        '200':
          description: List of calendars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Calendar'
        '401':
          description: Unauthorized
        '500':
          description: Server error

    post:
      summary: Create a new calendar
      tags:
        - Calendars
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                groupId:
                  type: string
                  format: uuid
                  description: Target group ID (must be a group the user belongs to). If omitted, uses the user's first group.
                name:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
              required:
                - name
                - color
      responses:
        '201':
          description: Calendar created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not a member of the specified group
        '400':
          description: Invalid input
        '500':
          description: Server error

  /calendars/{id}:
    patch:
      summary: Update calendar (name/color)
      tags:
        - Calendars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '200':
          description: Calendar updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '401':
          description: Unauthorized
        '404':
          description: Calendar not found
        '500':
          description: Server error

    delete:
      summary: Delete calendar
      tags:
        - Calendars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Calendar deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Calendar not found
        '500':
          description: Server error

  /events:
    get:
      summary: Get events for a specific month (across all groups the user belongs to)
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            format: 'YYYY-MM'
            example: '2026-06'
        - name: groupId
          in: query
          schema:
            type: string
            format: uuid
          description: Optional - filter by specific group
      responses:
        '200':
          description: List of events for the month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
        '400':
          description: Invalid month format
        '500':
          description: Server error

    post:
      summary: Create a new event (groupId is derived from the calendar)
      description: |
        イベントの `groupId` はリクエストで指定せず、`calendarId` からカレンダーを逆引きして自動設定される。
        フロントエンドは選択中のグループに属するカレンダーの `calendarId` を送ることで、正しいグループにイベントが作成される。
      tags:
        - Events
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                calendarId:
                  type: string
                  format: uuid
                  description: The calendar to add the event to. The event's groupId is automatically set to this calendar's groupId.
                title:
                  type: string
                memo:
                  type: string
                location:
                  type: string
                  maxLength: 200
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                startAt:
                  type: string
                  format: date-time
                endAt:
                  type: string
                  format: date-time
                isAllDay:
                  type: boolean
                  default: false
              required:
                - calendarId
                - title
                - startAt
                - endAt
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
        '400':
          description: Invalid input or calendar not found
        '403':
          description: Forbidden - not a member of the calendar's group
        '500':
          description: Server error

  /events/{id}:
    get:
      summary: Get an event by ID (any authenticated user)
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Event detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
        '404':
          description: Event not found
        '500':
          description: Server error

    patch:
      summary: Update event (creator only)
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                memo:
                  type: string
                location:
                  type: string
                  maxLength: 200
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                startAt:
                  type: string
                  format: date-time
                endAt:
                  type: string
                  format: date-time
                isAllDay:
                  type: boolean
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not event creator
        '404':
          description: Event not found
        '500':
          description: Server error

    delete:
      summary: Delete event (creator only)
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Event deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not event creator
        '404':
          description: Event not found
        '500':
          description: Server error

  /shift-patterns:
    get:
      summary: Get current user's shift patterns
      tags:
        - ShiftPatterns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of shift patterns sorted by sortOrder
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftPattern'
        '401':
          description: Unauthorized
        '500':
          description: Server error

    post:
      summary: Create a shift pattern
      tags:
        - ShiftPatterns
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                startTime:
                  type: string
                  example: '07:05'
                endTime:
                  type: string
                  example: '15:50'
                breakMinutes:
                  type: integer
                  default: 0
                isDayOff:
                  type: boolean
                  default: false
                sortOrder:
                  type: integer
                  default: 0
              required:
                - name
                - color
      responses:
        '201':
          description: Shift pattern created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftPattern'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /shift-patterns/{id}:
    get:
      summary: Get a shift pattern by ID (owner only)
      tags:
        - ShiftPatterns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Shift pattern detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftPattern'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not pattern owner
        '404':
          description: Shift pattern not found
        '500':
          description: Server error

    patch:
      summary: Update a shift pattern (owner only)
      tags:
        - ShiftPatterns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                startTime:
                  type: string
                endTime:
                  type: string
                breakMinutes:
                  type: integer
                isDayOff:
                  type: boolean
                sortOrder:
                  type: integer
      responses:
        '200':
          description: Shift pattern updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftPattern'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not pattern owner
        '404':
          description: Shift pattern not found
        '500':
          description: Server error

    delete:
      summary: Delete a shift pattern (owner only)
      tags:
        - ShiftPatterns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Shift pattern deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not pattern owner
        '404':
          description: Shift pattern not found
        '500':
          description: Server error

  /shifts:
    get:
      summary: Get shifts for all group members in a specific month (across all groups the user belongs to)
      tags:
        - Shifts
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            example: '2026-06'
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Optional - filter by specific group
      responses:
        '200':
          description: List of shifts for the month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '400':
          description: Invalid month format
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /shifts/{date}:
    put:
      summary: Assign or update the current user's shift on a specific date
      tags:
        - Shifts
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
            example: '2026-06-12'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shiftPatternId:
                  type: string
                  format: uuid
              required:
                - shiftPatternId
      responses:
        '200':
          description: Shift assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '401':
          description: Unauthorized
        '404':
          description: Shift pattern not found
        '500':
          description: Server error

    delete:
      summary: Remove the current user's shift on a specific date
      tags:
        - Shifts
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Shift removed
        '401':
          description: Unauthorized
        '404':
          description: Shift not found
        '500':
          description: Server error

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
        memo:
          type: string
          nullable: true
          maxLength: 200
          description: 自己紹介・ひとことメモ
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt

    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        inviteToken:
          type: string
        ownerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - inviteToken
        - ownerId
        - createdAt

    GroupWithMembers:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: '#/components/schemas/User'

    Calendar:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - name
        - color
        - createdBy
        - createdAt
        - updatedAt

    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        calendarId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        title:
          type: string
        memo:
          type: string
        location:
          type: string
          maxLength: 200
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        isAllDay:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - calendarId
        - createdBy
        - title
        - startAt
        - endAt
        - isAllDay
        - createdAt
        - updatedAt

    GroupPreview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        memberCount:
          type: integer
      required:
        - id
        - name
        - memberCount

    ShiftPattern:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        startTime:
          type: string
          example: '07:05'
          nullable: true
        endTime:
          type: string
          example: '15:50'
          nullable: true
        breakMinutes:
          type: integer
        isDayOff:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - color
        - isDayOff
        - sortOrder
        - createdAt
        - updatedAt

    Shift:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        shiftPatternId:
          type: string
          format: uuid
        shiftPattern:
          $ref: '#/components/schemas/ShiftPattern'
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - groupId
        - shiftPatternId
        - date
        - createdAt
        - updatedAt

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT